+++ Hector/Elements/Dipole.h
@@ -12,7 +12,7 @@ namespace Hector
     {
       public:
         /// (virtual) class constructor
-        Dipole( const Type& type, const std::string& name, float spos, float length, float mag_str ) :
+        Dipole( const Type& type, const std::string& name, double spos, double length, double mag_str ) :
           ElementBase( type, name, spos, length ) {
           setMagneticStrength( mag_str );
         }
@@ -23,7 +23,7 @@ namespace Hector
     {
       public:
         /// Class constructor
-        RectangularDipole( const std::string& name, float spos, float length, float mag_str ) :
+        RectangularDipole( const std::string& name, double spos, double length, double mag_str ) :
           Dipole( aRectangularDipole, name, spos, length, mag_str ) {}
 
         std::shared_ptr<ElementBase> clone() const override { return std::make_shared<RectangularDipole>( *this ); }
@@ -43,7 +43,7 @@ namespace Hector
          */
         /// \note Numerical sensitivity (~\f$10^{-8}\f$ relative precision on a 64-bit Intel machine) expected with \f$ \frac{r}{E_{\mathrm{b}}} \left(1-\cos{\theta}\right)\f$.
         ///  Using \f$ \cos{2x} = 1-2\sin^{2}{x} \f$ to transform this term (see the variable called "simp")
-        Matrix matrix( float, float mp = Parameters::get()->beamParticlesMass(), int qp = Parameters::get()->beamParticlesCharge() ) const override;
+        Matrix matrix( double, double mp = Parameters::get()->beamParticlesMass(), int qp = Parameters::get()->beamParticlesCharge() ) const override;
     };
 
     /// Sector dipole object builder
@@ -51,7 +51,7 @@ namespace Hector
     {
       public:
         /// Class constructor
-        SectorDipole( const std::string& name, float spos, float length, float mag_str ) :
+        SectorDipole( const std::string& name, double spos, double length, double mag_str ) :
           Dipole( aSectorDipole, name, spos, length, mag_str ) {}
 
         std::shared_ptr<ElementBase> clone() const override { return std::make_shared<SectorDipole>( *this ); }
@@ -82,7 +82,7 @@ namespace Hector
          * \f$
          * assuming \f$\theta = {L\over r}\f$, \f$ {1\over r} \equiv k =  k_{0} \cdot \frac{p_{0}}{p_{0} - \mathrm{d}p} \cdot \frac{q_{\mathrm{part}}}{q_{\mathrm{b}}} \f$
          */
-        Matrix matrix( float, float mp = Parameters::get()->beamParticlesMass(), int qp = Parameters::get()->beamParticlesCharge() ) const override;
+        Matrix matrix( double, double mp = Parameters::get()->beamParticlesMass(), int qp = Parameters::get()->beamParticlesCharge() ) const override;
     };
   }
 }
