+++ Hector/Elements/Drift.cc
@@ -4,22 +4,22 @@ namespace Hector
 {
   namespace Element
   {
-    Drift::Drift( const std::string& name, float spos, float length ) :
+    Drift::Drift( const std::string& name, double spos, double length ) :
       ElementBase( aDrift, name, spos, length )
     {}
 
-    Drift::Drift( const std::string& name, const Type& type, float spos, float length ) :
+    Drift::Drift( const std::string& name, const Type& type, double spos, double length ) :
       ElementBase( type, name, spos, length )
     {}
 
     Matrix
-    Drift::matrix( float, float, int ) const
+    Drift::matrix( double, double, int ) const
     {
       return genericMatrix( length_ );
     }
 
     Matrix
-    Drift::genericMatrix( float length )
+    Drift::genericMatrix( double length )
     {
       Matrix mat = DiagonalMatrix( 6, 1 );
       mat( 1, 2 ) = length;
