+++ Hector/Elements/Marker.h
@@ -12,7 +12,7 @@ namespace Hector
     {
       public:
         /// Class constructor
-        Marker( const std::string& name, float spos, float length = 0. ) :
+        Marker( const std::string& name, double spos, double length = 0. ) :
           Drift( name, spos, length ) {}
 
         std::shared_ptr<ElementBase> clone() const override { return std::make_shared<Marker>( *this ); }
